css: fix styling of sidebar
In most places in CSS, the sidebar is being referenced by using
the `aside.sidebar` selector. The sidebar is being declared with
its identifier set to "sidebar", not its class, which causes the
selectors to not match the sidebar at all. Adjust the sidebar to
instead have a class and adapt the single occasion of "#sidebar"
to use ".sidebar".